Move friend-graph related code into its own module. Wow so many different approaches.
