+++ Procfile
@@ -1,9 +1,9 @@
 friend_collector: python -m app.friend_collection.batch_per_thread
-bq_grapher: python -m app.workers.bq_grapher
-bq_list_grapher: python -m app.workers.bq_list_grapher
+bq_grapher: python -m app.friend_graphs.bq_grapher
+bq_list_grapher: python -m app.friend_graphs.bq_list_grapher
 pg_migrate: python -m app.models
 pg_pipeline_user_friends: python -m app.workers.pg_pipeline_user_friends
 pg_pipeline_user_details: python -m app.workers.pg_pipeline_user_details
 pg_pipeline_retweeter_details: python -m app.workers.pg_pipeline_retweeter_details
-psycopg_list_grapher: python -m app.workers.psycopg_list_grapher
+psycopg_list_grapher: python -m app.friend_graphs.psycopg_list_grapher
 graph_analyzer: python -m app.graph_analyzer
