+++ app/friend_graphs/base_grapher.py
