+++ app/friend_graphs/bq_custom_grapher.py
@@ -6,7 +6,7 @@ from memory_profiler import profile
 from dotenv import load_dotenv
 
 from app.workers import fmt_ts, fmt_n
-from app.workers.bq_grapher import BigQueryGrapher
+from app.friend_graphs.bq_grapher import BigQueryGrapher
 
 load_dotenv()
