+++ app/friend_graphs/bq_list_grapher.py
@@ -5,7 +5,7 @@ from networkx import DiGraph
 from memory_profiler import profile
 
 from app.workers import fmt_ts, fmt_n
-from app.workers.bq_grapher import BigQueryGrapher
+from app.friend_graphs.bq_grapher import BigQueryGrapher
 
 class BigQueryListGrapher(BigQueryGrapher):
