+++ app/friend_graphs/psycopg_batch_grapher.py
@@ -2,7 +2,7 @@
 from networkx import DiGraph
 from memory_profiler import profile
 
-from app.workers.psycopg_grapher import PsycopgGrapher
+from app.friend_graphs.psycopg_grapher import PsycopgGrapher
 
 class Grapher(PsycopgGrapher):
