+++ app/friend_graphs/psycopg_grapher.py
@@ -6,7 +6,7 @@ from memory_profiler import profile
 from app import APP_ENV
 from app.models import DATABASE_URL, USER_FRIENDS_TABLE_NAME
 from app.workers import DRY_RUN, BATCH_SIZE, USERS_LIMIT, fmt_ts, fmt_n
-from app.workers.base_grapher import BaseGrapher
+from app.friend_graphs.base_grapher import BaseGrapher
 
 class PsycopgGrapher(BaseGrapher):
     def __init__(self, dry_run=DRY_RUN, batch_size=BATCH_SIZE, users_limit=USERS_LIMIT,
